web/requirements.txt
Django==1.8.5
redis==2.10.3
django-redis==4.3.0
gunicorn==19.3.0
psycopg2==2.6
django-grappelli
django-debug-toolbar
Pillow
django-tastypie
requests